feat: Add version checking for updates
Introduces a component to check for application updates by comparing the locally defined `APP_VERSION` with the `version.json` file served by the application. If an update is available, a prompt will appear suggesting a reload. This ensures users are using the latest version of the Retro Invoice & Expense Tracker.